requirements.txt

aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.10.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.56.1
openpyxl==3.1.5
orjson==3.10.12
overpy==0.7
packaging==24.2
pandas==2.2.3
pillow==11.0.0
polyline==2.0.2
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
yarl==1.18.3
gunicorn